+++ examples/shared/src/examples/FillRasterLayer/QueryWithBounds.tsx
@@ -1,7 +1,7 @@
 import {
   Camera,
-  Layer,
   GeoJSONSource,
+  Layer,
   Map,
   type MapRef,
 } from "@maplibre/maplibre-react-native";
@@ -12,20 +12,7 @@ import newYorkCityDistrictsFeatureCollection from "@/assets/geojson/new-york-cit
 import { Bubble } from "@/components/Bubble";
 import { MAPLIBRE_DEMO_STYLE } from "@/constants/MAPLIBRE_DEMO_STYLE";
 
-const styles = {
-  neighborhoods: {
-    fillAntialias: true,
-    fillColor: "blue",
-    fillOutlineColor: "black",
-    fillOpacity: 0.84,
-  },
-  selectedNeighborhoods: {
-    fillAntialias: true,
-    fillColor: "green",
-    fillOpacity: 0.84,
-  },
-  bubbleText: { textAlign: "center" as const },
-};
+const LAYER = "fill";
 
 export function QueryWithBounds() {
   const mapRef = useRef<MapRef>(null);
@@ -46,55 +33,66 @@ export function QueryWithBounds() {
         ref={mapRef}
         mapStyle={MAPLIBRE_DEMO_STYLE}
         onPress={async (event) => {
-          const [longitude, latitude] = event.nativeEvent.lngLat;
-          const newBounds = [...(bounds ?? []), longitude, latitude];
-          if (newBounds.length === 4 && mapRef.current) {
-            const minX = Math.min(newBounds[0]!, newBounds[2]!);
-            const maxX = Math.max(newBounds[0]!, newBounds[2]!);
-            const minY = Math.min(newBounds[1]!, newBounds[3]!);
-            const maxY = Math.max(newBounds[1]!, newBounds[3]!);
+          const [x, y] = event.nativeEvent.point;
+
+          const pixelBounds = [...(bounds ?? []), x, y];
+          if (pixelBounds.length === 4 && mapRef.current) {
+            const minX = Math.min(pixelBounds[0]!, pixelBounds[2]!);
+            const minY = Math.min(pixelBounds[1]!, pixelBounds[3]!);
+            const maxX = Math.max(pixelBounds[0]!, pixelBounds[2]!);
+            const maxY = Math.max(pixelBounds[1]!, pixelBounds[3]!);
 
             const features = await mapRef.current.queryRenderedFeatures(
               [
                 [minX, minY],
                 [maxX, maxY],
               ],
-              { layers: ["nycFill"] },
+              { layers: [LAYER] },
             );
             setSelected(features);
             setBounds(undefined);
           } else {
-            setBounds(newBounds);
+            setBounds(pixelBounds);
           }
         }}
       >
         <Camera zoom={9} center={[-73.970895, 40.723279]} />
 
         <GeoJSONSource
-          id="nyc"
           data={
             newYorkCityDistrictsFeatureCollection as GeoJSON.FeatureCollection
           }
         >
-          <Layer type="fill" id="nycFill" style={styles.neighborhoods} />
+          <Layer
+            id={LAYER}
+            type="fill"
+            paint={{
+              "fill-antialias": true,
+              "fill-color": "blue",
+              "fill-outline-color": "black",
+              "fill-opacity": 0.84,
+            }}
+          />
         </GeoJSONSource>
 
         {selected ? (
           <GeoJSONSource
-            id="selectedNYC"
             data={{ type: "FeatureCollection", features: selected }}
           >
             <Layer
               type="fill"
-              id="selectedNYCFill"
-              style={styles.selectedNeighborhoods}
+              paint={{
+                "fill-antialias": true,
+                "fill-color": "green",
+                "fill-opacity": 0.84,
+              }}
             />
           </GeoJSONSource>
         ) : null}
       </Map>
 
       <Bubble>
-        <Text style={styles.bubbleText}>{message}</Text>
+        <Text style={{ textAlign: "center" }}>{message}</Text>
       </Bubble>
     </>
   );