+++ src/components/annotations/marker/Marker.tsx
@@ -1,7 +1,9 @@
 import {
-  type Component,
+  Component,
   type ComponentProps,
   type ReactElement,
+  type Ref,
+  useImperativeHandle,
   useRef,
 } from "react";
 import {
@@ -17,9 +19,26 @@ import { type Anchor, anchorToNative } from "../../../types/Anchor";
 import type { LngLat } from "../../../types/LngLat";
 import type { PixelPoint } from "../../../types/PixelPoint";
 import {
+  type NativeViewAnnotationRef,
   ViewAnnotation,
   type ViewAnnotationRef,
 } from "../view-annotation/ViewAnnotation";
+
+export type NativeMarkerRef = Component<
+  ComponentProps<typeof MarkerViewNativeComponent>
+> &
+  ReactNativeElement;
+
+export interface MarkerRef {
+  /**
+   * Returns a reference to the native component for Reanimated compatibility.
+   * This method is used by Reanimated's createAnimatedComponent to determine
+   * which component should receive animated props.
+   *
+   * @see https://docs.swmansion.com/react-native-reanimated/docs/core/createAnimatedComponent/#component
+   */
+  getAnimatableRef(): NativeMarkerRef | NativeViewAnnotationRef | null;
+}
 
 export interface MarkerProps extends ViewProps {
   /**
@@ -58,9 +77,14 @@ export interface MarkerProps extends ViewProps {
   selected?: boolean;
 
   /**
-   * Expects one child - can be container with multiple elements
+   * Expects one child - can be a View with multiple elements.
    */
   children: ReactElement;
+
+  /**
+   * Ref to access Marker methods.
+   */
+  ref?: Ref<MarkerRef>;
 }
 
 /**
@@ -76,18 +100,26 @@ export const Marker = ({
   id,
   anchor = "center",
   offset,
+  ref,
   ...props
 }: MarkerProps) => {
-  const nativeRef = useRef<
-    Component<ComponentProps<typeof MarkerViewNativeComponent>> &
-      ReactNativeElement
-  >(null);
+  const nativeRef = useRef<NativeMarkerRef>(null);
+  const viewAnnotationRef = useRef<ViewAnnotationRef>(null);
 
   const nativeAnchor = anchorToNative(anchor);
   const nativeOffset = offset ? { x: offset[0], y: offset[1] } : undefined;
 
-  const viewAnnotationRef = useRef<ViewAnnotationRef>(null);
   const frozenId = useFrozenId(id);
+
+  useImperativeHandle(
+    ref,
+    (): MarkerRef => ({
+      getAnimatableRef: () =>
+        Platform.OS === "ios"
+          ? (viewAnnotationRef.current?.getAnimatableRef() ?? null)
+          : nativeRef.current,
+    }),
+  );
 
   if (Platform.OS === "ios") {
     return (