+++ src/components/map/Map.tsx
@@ -142,18 +142,18 @@ export interface MapRef {
    * Converts geographic coordinates to pixel point of the view
    *
    * @example
-   * await mapRef.current?.getPointInView([-37.817070, 144.949901]);
+   * await mapRef.current?.project([13.04214014753952, 47.80554907882145]);
    *
-   * @param coordinates Geographic coordinate
+   * @param lngLat Geographic coordinate
    * @return Pixel point
    */
-  project(coordinates: LngLat): Promise<PixelPoint>;
+  project(lngLat: LngLat): Promise<PixelPoint>;
 
   /**
    * Converts a pixel point of the view to geographic coordinates.
    *
    * @example
-   * await mapRef.current?.getCoordinateFromView([280, 640]);
+   * await mapRef.current?.unproject([280, 640]);
    *
    * @param point Pixel point
    * @return Geographic coordinate